rockchip: rk3399: Add support for chromebook_scarlet
Add support for Scarlet, an RK3399-based table-formfactor chromebook in
the Gru board series.

This is only tested as a LEGACY_RW firmware payload.

The exact devicetree to use depends on the DSI panel in use, the
default is configured for the Kingdisplay panel variant, but the display
currently fails to initialize correctly.

With this Scarlet can boot into a usable serial u-boot prompt (requires
SuzyQ cable to access the AP console).